# Use a generic getter with __callee__