@headius 2023-08-09
Relates to https://github.com/jruby/jruby/issues/7876 where we noticed ostruct uses a LOT more memory now that it pre-creates all singleton methods in response to the hash form of `OpenStruct#new`.