@headius 2023-08-09
Note: I also tried using `class_eval` to create these methods with literal symbols. That was easily the fastest-performing mechanism for getter/setter methods, but it also consumed even more memory than the Proc version. 🤷‍♂️