@headius 2023-08-09
FWIW I know this library is not intended for performance-sensitive applications, but I think the speed and memory improvements here are worth the small change.